Rhythm.vextab was called two times because of resize detected. Fixed by checking in resize callback that width actually changed
Player.getNotes was called several time because of units that changed each time.
Fixed by letting Compiler.getRhythmUnit cache the result
This caused the player controls to be broken since the player instance had changed (old player kept playing in the background).
Now fixed but to avoid such a case, we stop previous player if any when creating a new instance.

General: artist and year as h3 instead of h2, tempo now displayed above rhythms.

Chords, Rhythm: Used a primereact Checkbox instead of a react-select Select for showing inline chords/rhythms.
Removed forced width of 250px which was a dirty fix for the old drop-zone on the right side preventing clicks.
Select will probably be removed, we are going to use primereact only for option forms (TODO: options in Ascii)