New component "Patchwork" which layouts its children according to props.layout. - props.layout must be the root node of a tree-like structure - each non leave node must have properties 'left' and 'right' or 'top' and 'bottom': child nodes (leave or non leave) - each leave node must be an array of the indices of component(s) in this.props.children - the root node can also be a leave node
Result will be a combination of SplitPane (horizontal or vertical) and Tabs to match the requested layout.
Tabs is used when several components are in a same leave node (i.e. when the leave node is an array containing more than one element).
When creating a Tabs, we use the child props.label as the tab title.

The last selected tab in each Tabs is remembered in Patchwork.state.tabIndex.
We use an Immutable Map, with keys being Immutable List (list of component indices) and values being the selected tab index (npm install immutable).
This means we store the selected tab index for a given combination of components.

Patchwork is used by App as a replacement for the existing combination of Tabs and SplitPane (removed these imports as well as state.tabIndex and TabsPane/EditorPane methods).
The props.layout of this Patchwork is currently driven by state.editorPosition, thus getting the same behavior as before.

Misc.:
- Used a distinct canvas id in Rhythm and Score so that we can display them at the same time (i.e. in the same DOM).
- Padding 0 10px is set in components General, Chords, Rhythm instead of the first column of the now removed div.columns2.
- Tempo is now displayed by the Rhythm component