Score: improvements regarding to the fact that drawing a score can take some seconds
1/ "Loading..." message (and hidden canvas) while the vextab is being (re)drawn.
ReactSizeDetector is added only after loading as it otherwise causes vextab to be called in a loop.
NB: I tried to use a fancy ProgressSpinner but it's not smooth at all (important jitter) while canvas is being drawn.

2/ When resizing, we apply a timeout of 50ms in order to keep resizing smooth (user has to stop dragging for 50ms for redraw to happen).
Loading message is not displayed (and canvas not hidden) when redrawing because of a resize.
Resize callback does nothing if size did not change.