https://reactjs.org/docs/refs-and-the-dom.html:
Caveats
If the ref callback is defined as an inline function, it will get called twice during updates, first with null and then again with the DOM element. This is because a new instance of the function is created with each render, so React needs to clear the old ref and set up the new one. You can avoid this by defining the ref callback as a bound method on the class, but note that it shouldn’t matter in most cases.

=> provided a bound method a ref + removed setTimeout in onChange